Disable auto-commit for JDBC connections
Execution of SQL interfaces aborts with an error when using
the Postgres JDBC driver in default auto-commit mode. This
is because DbImpl::executeImpl always attempts to commit an
open connection, irrespective if a transaction is open.

The error returned by the Postgres database is:

25P01 : no_active_sql_transaction

This patch disables auto-commit for all connections.

Unfortunately, it's not really a good solution, because disabling
auto-commit might cause unnecessary network round-trips for the
majority of single-statement queries.

A better solution would:

- execute single statement operations from SqlProxy honouring
  the auto-commit state of the connection
- execute Transaction operations by always disabling auto-commit
  mode on the top-level connection
- execute batch operations from SqlProxy in a mode explicitly
  set by the user

Also see https://github.com/mjdbc/mjdbc/pull/2